+++ RELEASENOTES.md
@@ -5,5 +5,5 @@
 - fds-funktionslink - fjernet asLink mulighed - op til bruger selv at udvide evt styling
 - fds-link WIP
 - DKFDS 8.1.0 rettelser
-  - Karakterbegrænsning, nyt i dkfds 8.1.0
+  - fds-input-limit - Karakterbegrænsning, nyt i dkfds 8.1.0
   - Sidenavigation omdøbt til venstremenu